Version2

+++ main.py
@@ -5,6 +5,10 @@ import webapp2
 
 from models import Message
 
+from google.appengine.api import users
+
+
+
 template_dir = os.path.join(os.path.dirname(__file__), "templates")
 jinja_env = jinja2.Environment(loader=jinja2.FileSystemLoader(template_dir), autoescape=False)
 
@@ -12,6 +16,7 @@ jinja_env = jinja2.Environment(loader=jinja2.FileSystemLoader(template_dir), aut
 class BaseHandler(webapp2.RequestHandler):
 
     def write(self, *a, **kw):
+        # type: (object, object) -> object
         return self.response.out.write(*a, **kw)
 
     def render_str(self, template, **params):
@@ -30,13 +35,26 @@ class BaseHandler(webapp2.RequestHandler):
 
 class MainHandler(BaseHandler):
     def get(self):
-        return self.render_template("hello.html")
+
+        user = users.get_current_user()
+        if user:
+            logout_url = users.create_logout_url("/")
+            logiran = True
+            podatki = {"logout_url": logout_url, "logiran": logiran, "user": user}
+        else:
+            login_url = users.create_login_url("/")
+            logiran = False
+            podatki = {"login_url": login_url, "logiran": logiran}
+
+        return self.render_template("hello.html", params=podatki)
 
 class ResultHandler(BaseHandler):
     def post(self):
+        user = users.get_current_user()
         first_name = self.request.get("first_name")
         last_name = self.request.get("last_name")
-        email = self.request.get("email")
+        """email = self.request.get("email")"""
+        email = user.email()
         message = self.request.get("message")
 
         result = Message(first_name=first_name, last_name=last_name, email=email, message=message)
@@ -47,17 +65,23 @@ class ResultHandler(BaseHandler):
 class ListHandler(BaseHandler):
     def get(self):
         list_ = Message.query(Message.deleted == False).fetch()
-        params = {"list_": list_}
+        user = users.get_current_user()
+        user_email = user.email()
+        params = {"list_": list_, "user_email": user_email}
         return self.render_template("list.html", params=params)
 
 class ListEditHandler(BaseHandler):
     def get(self, message_id):
+        if not users.is_current_user_admin():
+            return self.render_template("not_admin.html")
         result = Message.get_by_id(int(message_id))
 
         params = {"result": result}
         return self.render_template("list_edit.html", params=params)
 
     def post(self, message_id):
+        if not users.is_current_user_admin():
+            return self.render_template("not_admin.html")
         result = Message.get_by_id(int(message_id))
         result.message = self.request.get("new_text")
         result.put()
@@ -65,12 +89,16 @@ class ListEditHandler(BaseHandler):
 
 class ListDeleteHandler(BaseHandler):
     def get(self, message_id):
+        if not users.is_current_user_admin():
+            return self.render_template("not_admin.html")
         result = Message.get_by_id(int(message_id))
 
         params = {"result": result}
         return self.render_template("list_delete.html", params=params)
 
     def post(self, message_id):
+        if not users.is_current_user_admin():
+            return self.render_template("not_admin.html")
         result = Message.get_by_id(int(message_id))
         result.deleted = True
         result.put()
@@ -78,18 +106,24 @@ class ListDeleteHandler(BaseHandler):
 
 class DeletedListHandler(BaseHandler):
     def get(self):
+        if not users.is_current_user_admin():
+            return self.render_template("not_admin.html")
         list_ = Message.query(Message.deleted == True).fetch()
         params = {"list_": list_}
         return self.render_template("deleted_list.html", params=params)
 
 class DeletedListRecoverHandler(BaseHandler):
     def get(self, message_id):
+        if not users.is_current_user_admin():
+            return self.render_template("not_admin.html")
         result = Message.get_by_id(int(message_id))
 
         params = {"result": result}
         return self.render_template("list_recover.html", params=params)
 
     def post(self, message_id):
+        if not users.is_current_user_admin():
+            return self.render_template("not_admin.html")
         result = Message.get_by_id(int(message_id))
         result.deleted = False
         result.put()
@@ -97,12 +131,16 @@ class DeletedListRecoverHandler(BaseHandler):
 
 class DeletedListDeleteHandler(BaseHandler):
     def get(self, message_id):
+        if not users.is_current_user_admin():
+            return self.render_template("not_admin.html")
         result = Message.get_by_id(int(message_id))
 
         params = {"result": result}
         return self.render_template("list_delete_final.html", params=params)
 
     def post(self, message_id):
+        if not users.is_current_user_admin():
+            return self.render_template("not_admin.html")
         result = Message.get_by_id(int(message_id))
         result.key.delete()
         return self.redirect_to("deleted_list")
